Uplodaded all the files in 1 commit
It's a past project for my university class where needed to create an "OOP" Python software with GUI to enter and modify some data shown in the picture. It imports the data from "/data_json/data.json" and writes it back to "/data_json/data_out.json" when editing it on the GUI. Created as an "AI picture processing corrector" applicaiton.